+++ colormaps/src/main/java/colormaps/impl/ConstantGreen.java
@@ -26,9 +26,9 @@ import colormaps.ColorSpace;
 public class ConstantGreen extends AbstractColormap2D {
 
 	@Override
-	public Color getColor(float x, float y) {
+	public Color getColor(double x, double y) {
 		checkRanges(x, y);
-		return new Color(x, 0.5f, y);
+		return new Color((float)x, 0.5f, (float)y);
 	}
 
 	@Override
